+++ app/Http/Controllers/AdminOptionsController.php
@@ -2,6 +2,7 @@
 
 namespace App\Http\Controllers;
 
+use Illuminate\Support\Facades\Gate;
 use Illuminate\Http\Request;
 use App\Theme;
 use App\Option;
@@ -15,22 +16,38 @@ class AdminOptionsController extends Controller
 
     public function create(Theme $theme)
     {
+        if (Gate::denies('admin-access', User::class)) {
+            return 'Access denied';
+        }
+
         return view('adminOptions.create', compact('theme'));
     }
 
     public function store(Request $request, Theme $theme)
     {
+        if (Gate::denies('admin-access', User::class)) {
+            return 'Access denied';
+        }
+
         $theme->addOption($request);
         return redirect()->route('admin.themes.show')->with('flash_message', 'New option created');
     }
 
     public function edit(Option $option)
     {
+        if (Gate::denies('admin-access', User::class)) {
+            return 'Access denied';
+        }
+
         return view('adminOptions.edit', compact('option'));
     }
 
     public function update(Request $request, Option $option)
     {
+        if (Gate::denies('admin-access', User::class)) {
+            return 'Access denied';
+        }
+
         $option->update($request->all());
         return redirect()->route('admin.themes.show')->with('flash_message', 'Option updated');
 
@@ -38,10 +55,12 @@ class AdminOptionsController extends Controller
 
     public function destroy(Request $request, Option $option)
     {
+        if (Gate::denies('admin-access', User::class)) {
+            return 'Access denied';
+        }
+
         $option->delete();
-
         session()->flash('flash_message', 'Option deleted');
-
         return back();
 
     }
